LI12223

Description

Tipos de dados e funções auxiliares para a realização do projeto de LI1 em 2022/23.

Tipos de dados

Básicos

type Coordenadas = (Int, Int)

Par de coordenadas de uma posição no Mapa.

type Largura = Int

Comprimento de um Mapa.

type Velocidade = Int

Velocidade que irá afetar a movimentação dos Obstaculos de um Mapa.

Mapas

data Mapa

O Mapa que constituí o Jogo.

Constructors

Mapa Largura [(Terreno, [Obstaculo])]

data Terreno

Cada linha de um Mapa é um Terreno em particular contendo Obstaculos.

As linhas do tipo Rio ou Estrada têm a propriedade Velocidade que indicam a velocidade de deslocamento dos obstáculos. Podendo esta ser negativa, indicando assim a sua direção.

Constructors

Rio Velocidade

Estrada Velocidade

Relva

data Obstaculo

Um Obstáculo numa linha de um Mapa.

Constructors

Nenhum
a ausência de obstáculos

Tronco
os troncos deslizam apenas em Rio

Carro
os carros movimentam-se apenas em Estrada

Arvore
as árvores são um obstáculo fixo que não se move e apenas são possíveis em Relva

Jogo

data Jogo

Definição base de um jogo.

Constructors

Jogo

Fields

Jogador

o personagem do jogo

Mapa

o mapa em que se está a jogar

newtype Jogador

O Jogador define o personagem controlado no Jogo.

Constructors

Jogador Coordenadas

data Direcao

Direção de uma Jogada feita por um Jogador no Mapa.

Constructors

Cima

Baixo

Esquerda

Direita

data Jogada

As acções que podem ser tomadas pelo Jogador em cada estado do Jogo.

Constructors

Parado
tipo que define a ausência de uma acção do Jogador

Move Direcao
um movimento do jogador numa determinada Direcao