Tarefa2_2022li1g002

Description

Módulo para a realização da Tarefa 2 do projeto de LI1 em 2022/23.

Documentation

estendeMapa

Arguments

:: Mapa
Mapa antecedente da nova linha.

-> Int
Seed.

-> Mapa
Mapa com a nova linha.

A função não recursiva estendeMapa recebe um mapa e uma seed (um número de 1 a 100 usado para gerar os números aleatórios).

Funcionamento da aleatoriedade:

Através da seed recebida como argumento, a função obterRandoms gera uma lista com números de 19 dígitos. O tamanho desta lista corresponde à largura do mapa mais uma unidade. A head da lista é usada para gerar o terreno, já os restantes elementos são utilizados para gerar os obstáculos.

Exemplo de utilização:

>>> estendeMapa (Mapa 5 [(Relva, [Arvore, Nenhum, Arvore, Nenhum, Arvore]),(Estrada (-1), [Nenhum, Nenhum, Nenhum, Carro, Carro]),(Relva, [Arvore, Nenhum, Nenhum, Arvore, Arvore])])
Mapa 5 [(Relva, [Arvore, Nenhum, Arvore, Nenhum, Arvore]),(Estrada (-1), [Nenhum, Nenhum, Nenhum, Carro, Carro]),(Relva, [Arvore, Nenhum, Nenhum, Arvore, Arvore]),(Rio, [Nenhum, Tronco, Tronco, Nenhum, Tronco])]

gerarObstaculos

Arguments

:: [Int]
Lista de números gerados aleatoriamente.

-> Int
Largura do mapa que irá ser prolongado.

-> Terreno
Tipo de terreno para o qual se gerará obstáculos.

-> [Obstaculo]
Lista com os tipos de obstáculos possíveis para um certo tipo de terreno.

-> [Obstaculo]
Obstáculos gerados específicos para a nova linha.

A função recursiva gerarObstaculos recebe a lista de inteiros criada pela função auxiliar obterRandoms, a largura do mapa, o terreno da nova linha e os possíveis obstáculos para o mesmo. A função irá devolver uma lista com todos os obstáculos a introduzir nessa nova linha, de um certo tipo de terreno.

gerarTerreno

Arguments

:: Int
Primeiro número gerado que irá ser aplicado para descobrir qual o tipo de terreno escolhido aleatoriamente.

-> [Terreno]
Lista com os tipos de terreno possíveis da qual se escolherá um aleatoriamente.

-> Terreno
Terreno escolhido aleatoriamente.

A função não recursiva gerarTerreno recebe um inteiro, a cabeça da lista de inteiros gerados aleatoriamente pela função auxiliar obterRandoms, e ainda uma lista com todos os tipos de terreno possíveis. Assim, a função irá devolver um tipo de terreno, o qual será aplicado na nova linha a adicionar ao mapa inicial.

obterRandoms

Arguments

:: Int
Seed.

-> Int
Número inteiro obtido através da soma de 1 unidade à largura do mapa a prolongar.

-> [Int]
Lista de números gerados aleatoriamente.

A função não recursiva obterRandoms recebe um inteiro, a Seed, e outro inteiro, a soma de 1 unidade ao valor da largura do mapa a prolongar. Com estes valores, a função ao ser executada irá devolver uma lista de inteiros gerados aleatoriamente, que serão utilizados nas funções gerarTerreno e gerarObstaculos.

proximosTerrenosValidos

Arguments

:: Mapa
Mapa antecedente à nova linha, ou seja, a prolongar.

-> Int
Número pseudo aleatório que vai ser usado para gerar uma velocidade (para terrenos rio e estrada).

-> [Terreno]
Lista de terrenos válidos que podem pertencer à nova linha a adicionar.

A função não recursiva proximosTerrenosValidos recebe o mapa a prolongar e irá devolver uma lista de terrenos válidos para a nova linha, tendo já sido previamente aplicadas as condições necessárias do jogo, analisadas pelas funções auxiliares rioAux, estradaAux e relvaAux.

Exemplo de utilização:

>>> proximosTerrenosValidos (Mapa 5 [(Relva, [Arvore, Nenhum, Arvore, Nenhum, Arvore]),(Estrada (-1), [Nenhum, Nenhum, Nenhum, Carro, Carro]),(Relva, [Arvore, Nenhum, Nenhum, Arvore, Arvore])])
[Rio 0, Estrada 0, Relva]

geraVelocidade :: Int -> Velocidade

A função não recursiva geraVelocidade é um algoritmo pseudo-aleatório para gerar a velocidade de um rio ou estrada.

rioAux

Arguments

:: Mapa
Mapa antecedente à nova linha, ou seja, a prolongar.

-> Int
Contador.

-> Velocidade
Velocidade caso seja necessária.

-> [Terreno]
Lista de terrenos válidos que podem pertencer à nova linha a adicionar quando a última linha tem como terreno um rio.

A função recursiva rioAux recebe o mapa a prolongar e um inteiro, que irá funcionar como contador, devolvendo a lista de terrenos válidos para aplicar na nova linha, tendo em conta que a linha anterior é um rio e os tipos de terreno das linhas que precedem esta.

estradaAux

Arguments

:: Mapa
Mapa antecedente à nova linha, ou seja, a prolongar.

-> Int
Contador.

-> Velocidade
Velocidade caso seja necessária.

-> [Terreno]
Lista de terrenos válidos que podem pertencer à nova linha a adicionar quando a última linha tem como terreno uma estrada.

A função recursiva estradaAux recebe o mapa a prolongar e um inteiro, que irá funcionar como contador, devolvendo a lista de terrenos válidos para aplicar na nova linha, tendo em conta que a linha anterior é uma estrada e os tipos de terreno das linhas que precedem esta.

relvaAux

Arguments

:: Mapa
Mapa antecedente à nova linha, ou seja, a prolongar.

-> Int
Contador.

-> Velocidade
Velocidade caso seja necessária.

-> [Terreno]
Lista de terrenos válidos que podem pertencer à nova linha a adicionar quando a última linha tem como terreno relva.

A função recursiva relvaAux recebe o mapa a prolongar e um inteiro, que irá funcionar como contador, devolvendo a lista de terrenos válidos para aplicar na nova linha, tendo em conta que a linha anterior é relva e os tipos de terreno das linhas que precedem esta.

proximosObstaculosValidos

Arguments

:: Int
Largura do mapa a prolongar.

-> (Terreno, [Obstaculo])
Par constituído pelo tipo de terreno e por lista de obstáculos que descreve uma linha do mapa.

-> [Obstaculo]
Lista de obstáculos válidos para a nova linha a ser adicionada.

A função não recursiva proximosObstaculosValidos recebe a largura do mapa a prolongar e uma linha do mapa, devolvendo uma lista de obstáculos possíveis para preencher a lista dos mesmos da nova linha tendo em conta o tipo de terreno da mesma, utilizando as funções auxiliares proximosObstaculosEstradaAux e proximosObstaculosRioAux.

Exemplo de utilização:

>>> proximosObstaculosValidos 5 (Relva, [Arvore, Nenhum, Arvore, Nenhum])
[Nenhum, Arvore]

proximosObstaculosEstradaAux

Arguments

:: [Obstaculo]
Lista de obstáculos de uma linha do mapa a prolongar se o terreno for estrada.

-> Int
Contador.

-> [Obstaculo]
Lista de obstáculos possíveis para uma estrada verificando as condições necessárias exigidas.

A função recursiva proximosObstaculosEstradaAux recebe uma lista de obstáculos e um inteiro, que irá funcionar como contador, devolvendo uma lista de obstáculos possíveis de acrescentar tendo em conta os obstáculos anteriores e que o terreno é uma estrada.

proximosObstaculosRioAux

Arguments

:: [Obstaculo]
Lista de obstáculos de uma linha do mapa a prolongar se o terreno for rio.

-> Int
Contador.

-> [Obstaculo]
Lista de obstáculos possíveis para um rio verificando as condições necessárias exigidas.

A função recursiva proximosObstaculosEstradaAux recebe uma lista de obstáculos e um inteiro, que irá funcionar como contador, devolvendo uma lista de obstáculos possíveis de acrescentar tendo em conta os obstáculos anteriores e que o seu terreno será um rio.

velocidadeMaxMin :: (Velocidade, Velocidade)