Tarefa3_2022li1g002

Description

Módulo para a realização da Tarefa 3 do projeto de LI1 em 2022/23.

Documentation

animaJogo

Arguments

:: Jogo
Estado do jogo antes da jogada.

-> Jogada
A jogada em si, Nenhum ou Move Direcao.

-> Jogo
Jogo após a jogada dada.

A função não recursiva animaJogo movimenta os obstáculos (de acordo com a velocidade do terreno em que se encontram) e a personagem, de acordo com a jogada dada.

Exemplo de utilização:

>>> animaJogo (Jogo (Jogador (2,2)) (Mapa 5 [(Relva, [Arvore, Nenhum, Arvore, Nenhum, Arvore]),(Estrada (-1), [Nenhum, Nenhum, Nenhum, Carro, Carro]),(Relva, [Arvore, Nenhum, Nenhum, Arvore, Arvore])])) (Move Esquerda)
Jogo (Jogador (1,2)) (Mapa 5 [(Relva,[Arvore,Nenhum,Arvore,Nenhum,Arvore]),(Estrada (-1),[Nenhum,Nenhum,Carro,Carro,Nenhum]),(Relva,[Arvore,Nenhum,Nenhum,Arvore,Arvore])])

moveJogador

Arguments

:: Jogador
Estado do jogador antes da jogada.

-> Jogada
A jogada em si, Nenhum ou Move Direcao.

-> [(Terreno, [Obstaculo])]
As linhas do mapa.

-> Jogador
Estado do jogador pós jogada.

A função não recursiva moveJogador recebe Jogador com as Coordenadas antes das execução da função, uma Jogada e a lista de linhas do mapa. Para o jogador se mover, têm de ser respeitadas as seguintes condições:

O jogador não pode sair do mapa (exceto quando está num tronco)

O jogador não pode ir para "cima" de árvores, ou seja, quando há uma árvore no caminho do jogador este fica parado no lugar inicial

Quando a jogada é Parado e o jogador se encontra na posição do tronco, este move-se junto com o tronco (|v| unidades na direção do rio)

Quando o jogador se move horizontalmente e se encontra num tronco, a sua posição resultante em x será a soma do vetor v com (x ± 1)

Exemplo de utilização:

>>> moveJogador (Jogador (1,1)) (Move Cima) [(Relva, [Arvore, Nenhum, Arvore]),(Relva, [Nenhum,Nenhum,Nenhum])]
Jogador (1,0)

>>> moveJogador (Jogador (0,1)) (Move Cima) [(Relva, [Arvore, Nenhum, Arvore]),(Relva, [Nenhum,Nenhum,Nenhum])]
Jogador (0,1)

animaObstaculos

Arguments

:: [(Terreno, [Obstaculo])]
Linhas do mapa antes da animação.

-> [(Terreno, [Obstaculo])]
Linhas do mapa depois do movimento dos obstáculos.

A função recursiva animaObstaculos recebe uma lista de linhas do mapa e, com ajuda da função auxiliar animaObstaculoAux e através da velocidade v do terreno, move todos os elementos da lista de obstáculos |v| unidades na direção da velocidade.

Exemplo de utilização:

>>> animaObstaculos [(Rio (-1), [Nenhum,Tronco,Tronco,Nenhum]),(Estrada 2, [Carro,Carro,Nenhum,Nenhum]),(Rio 1, [Tronco,Tronco,Nenhum,Nenhum])]
[(Rio (-1),[Tronco,Tronco,Nenhum,Nenhum]),(Estrada 2,[Nenhum,Nenhum,Carro,Carro]),(Rio 1,[Nenhum,Tronco,Tronco,Nenhum])]

animaObstaculoAux

Arguments

:: Velocidade
Velocidade do terreno a ser animado.

-> [Obstaculo]
Lista de obstáculos a ser movimentados.

-> Int
Contador.

-> [Obstaculo]
Obstáculos pós transformação no x.

A função auxiliar recursiva animaObstaculoAux recebe uma velocidade e uma lista de obstáculos, com esses valores "desvia" os elementos da lista pelo valor da velocidade.

Exemplo de utilização:

>>> animaObstaculoAux (-1) [Nenhum,Tronco,Tronco,Nenhum] 0
[Tronco,Tronco,Nenhum,Nenhum]