Tarefa5_2022li1g002

Description

Módulo para a realização da Tarefa 5 do projeto de LI1 em 2022/23.

Documentation

deslizaJogo

Arguments

:: Int
Seed.

-> Jogo
Jogo que irá ser deslizado.

-> Jogo
Jogo depois de ser deslizado, retirando a linha mais abaixo e adicionando uma nova no topo.

A função deslizaJogo permite que o jogo vá avançando a nível de linhas, sendo o mapa renovado, eliminando a linha mais abaixo e gerando uma nova no topo do mapa, sucessivamente.