Tarefa6_2022li1g002

Description

Módulo para a realização da Tarefa 6 do projeto de LI1 em 2022/23.

Documentation

data Cena

O tipo de dados Cena é usado no estado para controlar a lógica da aplicação, o grupo decidiu dividir o jogo em cenas para distinguir menus/modos de jogo.

Constructors

MainMenu Int
Menu principal do jogo.

JogoCena (Int, Int)
Jogo, o tuplo de Int serve para controlar os submenus, primeiro argumento: (0 -> Sem Menu, 1 -> Menu de Pausa, 2 -> Menu de Game Over), segundo argumento indica a opção do submenu selecionada.

Editor
Editor de mapas.

SkinsMenu
Menu de escolha de personagem.

data Estado

O tipo de dados Estado contém algumas variáveis globais usadas para atualizar o jogo, a função tempoReage utiliza este estado com frequência relativa aos fps do jogo. Este tipo de dados usa "record syntax" que permite ultilizar as suas variáveis pelo nome.

Constructors

Estado

Fields

debug :: Bool

O Bool debug serve para apresentar alguns dados na tela úteis para fins de desenvolvimento.

cena :: Cena

A Cena em que a aplicação se encontra.

jogo :: Jogo

Quando a aplicação se econtra em JogoCena esta variável com tipo definido em LI12223 guarda o estado do jogo estilo Crossy Road.

movimento :: Jogada

Quando o jogador prime uma tecla de movimento (setas), esta variável é alterada para indicar o movimento que o jogador deseja efetuar.

direcao :: Direcao

Esta variável serve para desenhar o jogador na tela tendo em conta a direção para onde "olha", normalmente corresponde à direção do último movimento efetuado.

vivo :: Bool

Indica se o jogo ainda não terminou.

tick :: Int

Para suavizar e animar o jogo independentemente dos fps utilizamos esta variável que é atualizada (+1) atualização do jogo, atualizamos o jogo quando fps == taxaUpdate.

seed :: Int

Seed do mapa (valor de 1 a 100 inclusive).

pontuacao :: (Int, Int)

Primeiro argumento do tuplo é a pontuação apresentada na tela, o segundo é uma variável de controlo (como o y não cresce à medida que o mapa avança é preciso saber quando é que o jogador realmente "sobe" uma unidade).

genY :: Int

Este valor é usado para gerar a próxima linha do mapa de forma dinâmica.

editor :: EstadoEditor

O EstadoEditor é usado na cena Editor.

skin :: Int

Indica a personagem que está a ser utilizada.

alturaMapa :: Int

Quantidade de linhas que aparecem na tela.

larguraMapa :: Int

Quantidade de colunas que aparecem na tela.

estadoInicial :: IO Estado

Estado inicial do jogo.

dm :: Display

Define o display da aplicação Gloss.

dimensoesDisplay :: (Int, Int)

Dimensões da janela do jogo (1600, 900).

imagensCaminho :: String

Caminho para os assets gráficos.

fps :: Int

Frames per second , número de vezes que o jogo é desenhado num segundo.

taxaUpdate :: Int

Número de ticks a que o jogo tem de chegar para voltar a ser atualizado.

drawEstado

Arguments

:: [[Picture]]
Lista de imagens previamente carregadas.

-> Estado
Estado da aplicação.

-> IO Picture
Retorna a Picture que neste caso é o frame completo que a aplicação desenhou.

A função drawEstado controla o que deve ser desenhado dependendo do estado da aplicação.

drawSkinsMenu

Arguments

:: Picture
Imagem correspondente ao fundo do menu.

-> [Picture]
Imagens de todas as personagens.

-> Int
Personagem selecionada.

-> Picture
Retorna a imagem gerada juntando os elementos que recebe.

A função drawSkinsMenu desenha o menu de escolha de personagem.

drawMenu

Arguments

:: Cena
A variável do tipo Cena contém a opção selecionada.

-> [Picture]
Imagens que formam o menu.

-> Picture
Retorna a imagem relativa à opção selecionada.

A função drawMenu desenha o menu principal.

drawJogoUI

Arguments

:: Int
Pontuação atual.

-> (Int, Int)
Submenu aberto e a sua opção selecionada.

-> [Picture]
Imagens do menu de pausa.

-> [Picture]
Imagens dos números que formam o mostrador da pontuação.

-> [Picture]
Imagens da tela de game over.

-> Picture
Imagem gerada.

A função drawJogoUI desenha a interface do jogo, pontuação, menu de pausa, tela de game over.

drawPontuacao

Arguments

:: Int
Número a ser desenhado.

-> [Picture]
Imagens dos números.

-> Picture
Número em imagem.

A função drawPontuacao é auxilar de drawJogoUI, desenha um n do tipo Int desde que n >= 0 && n < 1000 através de imagens com os números de 0 - 9 .

drawDebugUI

Arguments

:: Int
Seed do mapa atual.

-> (Int, Int)
Posição do jogador.

-> Int
Número de ticks desde a última atualização do mapa.

-> Direcao
Direção atual do jogador.

-> Bool
Direção atual do jogador.

-> Int
Bool que indica se o jogo terminou.

-> Picture
Retorna um painel de estatísticas.

A função drawDebugUI desenha algumas estatísticas para fins de desenvolvimento e debugging no canto superior esquerdo da aplicação.

drawMuros :: Picture

A função drawMuros desenha umas barreiras à volta do mapa para esconder alguns obstáculos que ao serem animados passam pelas bordas do mapa.

drawJogo

Arguments

:: [[Picture]]
Lista de imagens do jogo.

-> Mapa
Mapa a ser desenhado.

-> Coordenadas
Coordenadas da posição atual do jogador.

-> Direcao
Direção atual do jogador.

-> Int
Ticks desde a última atualização do jogo, usado para animar o os obstáculos suavemente (translação suave).

-> Bool
Indica se o mapa já deve deslizar.

-> Int
Personagem a ser utilizada pelo jogador.

-> Picture
Retorna a imagem do jogo.

A função drawJogo desenha o jogo principal da aplicação, utiliza as funções auxiliares drawJogador, drawMapa e drawObstaculos. É também encarregue do deslize suave do mapa.

drawJogador

Arguments

:: [Picture]
Lista de imagens do jogador (inclui todas as personagens).

-> Direcao
Direção atual do jogador.

-> Coordenadas
Coordenadas da posição atual do jogador.

-> Velocidade
Velocidade da linha onde o jogador se encontra (usado para animar o jogador suavemente quando este está em cima de um tronco).

-> Int
Ticks desde a última atualização do jogo.

-> Int
Personagem a ser utilizada.

-> Picture
Retorna a imagem do jogador.

A função drawJogador desenha o jogador na tela.

drawMapa

Arguments

:: [Picture]
Imagens dos terrenos.

-> Mapa
Mapa a ser desenhado.

-> (Float, Float)
(x,y) variam a cada iteração da função para desenhar o mapa de forma isométrica.

-> [Picture]
Lista de imagens geradas (uma para cada linha do mapa).

A função recursiva drawMapa desenha a base das linhas do jogo (terrenos) tendo em conta que o jogo é isométrico.

drawLinha

Arguments

:: [Picture]
Imagens dos terrenos.

-> Largura
Número de colunas a serem desenhadas.

-> (Terreno, [Obstaculo])
Linha a ser desenhada.

-> (Float, Float)
(x,y) variam a cada iteração da função para desenhar a linha de forma isométrica.

-> [Picture]
Linha em imagem.

A função recursiva drawLinha é auxiliar de drawMapa desenha uma linha na tela.

drawSegmento

Arguments

:: Picture
Asset da imagem a ser desenhada.

-> (Float, Float)
Posição da imagem na tela.

-> Picture
Imagem desenhada.

A função drawSegmento desenha uma imagem na grelha do jogo.

drawObstaculos

Arguments

:: [Picture]
Assets das imagens dos obstáculos.

-> Mapa
Mapa a ser desenhado.

-> (Float, Float)
(x,y) variam a cada iteração da função para desenhar a linha de forma isométrica.

-> Int
Ticks desde a última atualização do jogo.

-> Coordenadas
Posição do jogador no mapa, permite parar a animação dos carros no caso de atropelamento.

-> [Picture]
Lista de imagens geradas (uma para cada linha de obstáculos do mapa).

A função drawObstaculos desenha os obstáculos do jogo.

drawLinhaObstaculo

Arguments

:: [Picture]
Assets das imagens dos obstáculos.

-> [Obstaculo]
Lista de obstáculos a serem desenhados.

-> (Float, Float)
(x,y) variam a cada iteração da função para desenhar a linha de obstáculos de forma isométrica.

-> Velocidade
Velocidade da linha de obstáculos a ser desenhada.

-> Int
Ticks desde a última atualização do jogo.

-> Bool
Caso a linha tenha causado um atropelamento, este Bool indica se a animação dos carros deve parar.

-> [Picture]
Imagem da linha de obstáculos gerada.

A função drawLinhaObstaculo uma linha de obstáculos.

inputReage

Arguments

:: Event
Nesta aplicação só usamos eventos do tipo EventKey, numa futura autalização seria possível utilizar EventMotion para receber a posição do cursor e potencialmente permitir o uso do rato em menus.

-> Estado
Estado da aplicação.

-> IO Estado
Novo estado (pós input).

A função inputReage controla todas as operações que devem ser feitas quando a aplicação recebe um input do jogador.

proximoN

Arguments

:: Int
Opção do menu antes de receber o input.

-> Int
Número da última opção.

-> Int
(+1 ou -1) dependendo da seta em que o jogador carregou.

-> Int
Retorna a próxima opção.

A função proximoN é utilizada para decidir qual a próxima opção a ser selecionada quando o jogador clica nas setas num menu, evita bastante repetição de código.

tempoReage

Arguments

:: Float
Período de tempo em segundos que o jogo precisa de avançar (a nossa implementação não utiliza este argumento).

-> Estado
Estado antes da atualização.

-> IO Estado
Novo estado.

A função tempoReage é chamada sempre que um frame é desenhado, dependendo da cena, dos ticks e de outros argumentos do estado, a função decide que operação deve efetuar (se for necessário).

corFundo

Arguments

:: Color
Cor.

Cor do fundo da aplicação. (RGBA de 0 a 1)

main :: IO ()

A função main carrega todas as imagens para a memória da aplicação e chama a função playIO do Gloss. Para carregar imagens do tipo .png (ao contrário de .bmp), decidimos utlizar uma biblioteca externa Juicy.

geraMapa

Arguments

:: Largura
Largura do mapa a ser gerado.

-> Int
Altura do mapa.

-> Int
Seed.

-> Mapa
Mapa gerado.

A função não recursiva geraMapa, através da função auxiliar geraMapaAux e da função obterRandoms presente no módulo da tarefa 2, gera um mapa pseudo-aleatório utilizando uma seed e uma largura.

geraMapaAux

Arguments

:: [Int]
Lista de randoms.

-> Largura
Largura do mapa a ser gerado.

-> Mapa
Mapa gerado.

A função recursiva geraMapaAux cria linha a linha o mapa a ser gerado, tendo em conta que a primeira linha será relva, sem nenhum obstáculo.

arranjaRios

Arguments

:: [(Terreno, [Obstaculo])]
Linhas do mapa a serem corrigidas.

-> [(Terreno, [Obstaculo])]
Linhas pós correção dos rios.

Como na primeira parte do projeto o jogo foi feito a pensar no movimento do jogador sincronizado com a animação dos obstáculos, ao contrário da segunda parte (jogador move-se independente da animação dos obstáculos), precisamos de fazer uma correção ao movimento dos troncos para o jogador se conseguir mover.

data Chunk

Para o editor de mapas decidimos usar um tipo de dados diferente para a representação de linhas. Chunk pode ser Nada (sem terreno definido) ou '(Terreno, [Obstaculo])'

Constructors

Linha (Terreno, [Obstaculo])

Nada

type CameraEditor = (Int, Int, Float)

Estado da câmera do editor (x, y, zoom)

editorDeMapasImagens :: IO [[Picture]]

De forma a evitar o carregamento de imagens a mais no início da aplicação, o carregamento das imagens do editor faz-se apenas quando necessário.

alturaMapaEditor :: Int

Quantidade de linhas que aparecem na tela do editor.

larguraMapaEditor :: Int

Quantidade de colunas que aparecem na tela editor.

data EstadoEditor

O EstadoEditor controla os menus e interações do editor de mapas (faz parte do estado geral da aplicação mas só é alterado quando o utilizador está no editor de mapas).

Constructors

EstadoEditor

Fields

chunks :: [Chunk]

Lista de chunks do mapa a ser editado.

modo :: Int

Modo de edição.

camera :: CameraEditor

Definições da câmera.

toolbar :: Int

Toolbar a ser utilizada (toolbar de terrenos, toolbar de obstáculos).

itemSelecionado :: Int

Item da toolbar a ser utilizado.

objeto :: Coordenadas

Coordenadas do cursor do utilizador.

receberInput :: Bool

Indica se a aplicação está a receber inputs numa caixa de texto.

nomeFicheiro :: String

Nome do ficheiro (apenas utilizado quando o jogador está a gravar um mapa).

receberSelectMapa :: (Bool, Int)

Indica se a aplicação está a receber inputs numa caixa de seleção.

mapasGuardados :: [String]

Mapas gravados no diretório da aplicação.

estadoEditorInicial :: EstadoEditor

Estado inical do editor de mapas.

drawEditor

Arguments

:: EstadoEditor
Estado do editor de mapas a ser desenhado.

-> [[Picture]]
Imagens do terreno e obstáculos.

-> [[Picture]]
Imagens do UI do editor.

-> Picture
Retorna a imagem do estado do editor.

A função drawEditor desenha o editor de mapas.

drawUIEditor

Arguments

:: EstadoEditor
Estado do editor de mapas a ser desenhado.

-> [[Picture]]
Imagens do UI do editor.

-> Picture
Retorna a imagem da interface.

A função drawUIEditor desenha a interface do editor de mapas.

caixaInput

Arguments

:: String
String do que já foi escrito.

-> Picture
Imagem do fundo da caixa.

-> Picture
Retorna a imagem da caixa com o input dentro.

A função caixaInput desenha a interface da caixa de input de texto (quando o utilizador quer gravar um mapa tem de escrever o seu nome).

caixaSelect

Arguments

:: [String]
Nomes dos mapas no diretório da aplicação.

-> Int
Mapa selecionado.

-> Picture
Fundo da caixa de select.

-> Picture
Retorna a imagem da caixa com o nome dos mapas dentro.

A função caixaSelect desenha a interface da caixa de escolha (quando o utilizador quer escolher um mapa já feito).

formataSelects

Arguments

:: [Picture]
Textos com os nomes dos mapas.

-> Float
Y da label, diminui em cada iteração da função.

-> Int

-> [Picture]
Retorna os textos separados.

A função recursiva formataSelects é auxiliar de caixaSelect e espaça os nomes dos mapas uns dos outros na caixa.

desenhaCursor

Arguments

:: Coordenadas
Coordenadas do cursor.

-> Picture
Imagem do cursor.

-> Picture
Imagem do cursor posicionada na tela.

A função desenhaCursor desenha o cursor do utilizador no mapa.

drawChunks

Arguments

:: [Picture]
Imagens dos terrenos

-> [Picture]
Imagens dos obstáculos.

-> [Chunk]
Linhas do mapa a ser editado.

-> (Float, Float)
(x,y) variam a cada iteração da função para desenhar a linha de forma isométrica.

-> [Picture]
Imagens das linhas.

A função drawChunks desenha cada linha do mapa a ser editado.

drawLinhaNada

Arguments

:: [Picture]
Imagens dos terrenos.

-> Largura
Largura do mapa a ser desenhado.

-> (Float, Float)
(x,y) variam a cada iteração da função para desenhar a linha de forma isométrica.

-> [Picture]
Imagens da linha.

A função drawLinhaNada desenha linhas sem terreno.

editaChunk

Arguments

:: [Chunk]
Mapa a ser editado.

-> Coordenadas
Coordenadas onde o utilizador realizou a edição.

-> Int
Tipo de edição (alteração de terreno, adição de obstáuculo)

-> Int
ID do objeto.

-> [Chunk]
Mapa pós edição.

A função editaChunk recebe um Chunk e reliza a sua edição relativamente aos inputs do utilizador.

nomeMapasGuardados :: IO [String]

A função nomeMapasGuardados retorna os nomes de todos os mapas presentes no diretório do jogo.

leChunks

Arguments

:: String
Nome do ficheiro.

-> IO [Chunk]
Chunks carregados.

A função leChunks lê um ficheiro .mapa e faz parse de uma lista de [Chunk].

chunksParaMapa

Arguments

:: [Chunk]
Lista de Chunks a ser transformada.

-> Mapa
Mapa gerado.

A função chunksParaMapa tranforma uma lista de Chunks num Mapa (usado para o utilizador poder jogar o mapa que está a editar).

chunkParaLinha

Arguments

:: Chunk
Chunk a ser transformado.

-> (Terreno, [Obstaculo])
Linha gerada.

A função chunkParaLinha é auxiliar de chunksParaMapa, transforma um Chunk numa Linha do tipo (Terreno, [Obstaculo]).

chunksProntos

Arguments

:: [Chunk]
Lista de chunks.

-> Bool
Indica se a linha pode ser transformada.

A função chunksProntos indica se uma lista de chunks pode ser transformada numa lista de linhas (não pode ter linhas sem terreno).